refactor: add sync job for processing notifications
A note on the trigger and delay:
A defined trigger in the call to SyncNotifications means that we received
a push notification which has "triggered" this fetch. This also means
we had no active web socket and the app was in the background. As such,
it makes sense to provide a small delay in performing the sync so as
to wait for any other potential push notifications and batch the sync.