refactor: inline reporting of missing notifications
Since we now wait a few seconds before performing the fetch, it should
be fine to inline the asynchronous logic for reporting missing notifications.

This will now also mean the reporting is done within the bounds of the
fetch job and can't be dropped by the system.

Further more, we were not including the trigger notification as part of
the list of notifications to check, so this commit fixes that.

Finally, the ReceivedPushesEvent was removed, since this has proven to
be of little utility to us and in fact generates a lot of events.